:art: convert Layout to component
also some other migration stuff, and not importing unused code into components
