+++ gatsby-ssr.js
@@ -1,4 +1,4 @@
-import React from 'react'
+const React = require('react')
 
 exports.onRenderBody = ({ setPostBodyComponents }) => {
   const logger =
