+++ src/templates/blog-post.js
@@ -1,8 +1,9 @@
 import React from 'react'
+import { graphql } from 'gatsby'
 import Helmet from 'react-helmet'
-import { withComponent } from 'react-emotion'
 import format from 'date-fns/format'
 import addDays from 'date-fns/add_days'
+import Layout from '../components/Layout'
 import { Title, Subtitle, Rule } from '../components/Typography'
 import MarkdownContent from '../components/MarkdownContent'
 
@@ -11,13 +12,13 @@ import 'lowmess-prism'
 const Content = MarkdownContent.withComponent('main')
 const PostDate = Subtitle.withComponent('p')
 
-const BlogPostTemplate = ({ data }) => {
+const BlogPostTemplate = ({ location, data }) => {
   const post = data.markdownRemark
   // have to add a day because of timezone wonkiness
   const date = addDays(new Date(post.frontmatter.date), 1)
 
   return (
-    <article>
+    <Layout location={location}>
       <Helmet>
         <title>
           {post.frontmatter.title} • {data.site.siteMetadata.title}
@@ -41,33 +42,35 @@ const BlogPostTemplate = ({ data }) => {
           content={post.frontmatter.description}
         />
       </Helmet>
-      <header>
-        <Title
-          fontSize={[4, 5]}
-          fontWeight="7"
-          lineHeight="title"
-          mt={0}
-          mb={3}
-        >
-          {post.frontmatter.title}
-        </Title>
-        <PostDate
-          fontSize={[3, 4]}
-          fontWeight="5"
-          lineHeight="title"
-          mt={3}
-          mb={4}
-        >
-          <time dateTime={date}>{format(date, 'MMMM D, YYYY')}</time>
-        </PostDate>
-        <Rule mt={4} mb={5} />
-      </header>
-      <Content
-        lineHeight="copy"
-        fontSize={[1, 2]}
-        dangerouslySetInnerHTML={{ __html: post.html }}
-      />
-    </article>
+      <article>
+        <header>
+          <Title
+            fontSize={[4, 5]}
+            fontWeight="7"
+            lineHeight="title"
+            mt={0}
+            mb={3}
+          >
+            {post.frontmatter.title}
+          </Title>
+          <PostDate
+            fontSize={[3, 4]}
+            fontWeight="5"
+            lineHeight="title"
+            mt={3}
+            mb={4}
+          >
+            <time dateTime={date}>{format(date, 'MMMM D, YYYY')}</time>
+          </PostDate>
+          <Rule mt={4} mb={5} />
+        </header>
+        <Content
+          lineHeight="copy"
+          fontSize={[1, 2]}
+          dangerouslySetInnerHTML={{ __html: post.html }}
+        />
+      </article>
+    </Layout>
   )
 }
